+++ deploy/13-deploy-wstETH-weETH-oracles.ts
@@ -1,0 +1,68 @@
+import hre from "hardhat";
+import { DeployFunction } from "hardhat-deploy/dist/types";
+import { HardhatRuntimeEnvironment } from "hardhat/types";
+
+import { ADDRESSES } from "../helpers/deploymentConfig";
+
+const func: DeployFunction = async function ({
+  getNamedAccounts,
+  deployments,
+  network,
+  artifacts,
+}: HardhatRuntimeEnvironment) {
+  const { deploy } = deployments;
+  const { deployer } = await getNamedAccounts();
+  const proxyOwnerAddress = ADDRESSES[network.name].timelock;
+  const { WETH, wstETH, weETH } = ADDRESSES[network.name];
+
+  const resilientOracle = await hre.ethers.getContract("ResilientOracle");
+
+  let chainlinkOracle;
+  if (hre.network.name === "arbitrumone") {
+    chainlinkOracle = await hre.ethers.getContract("SequencerChainlinkOracle");
+  } else {
+    chainlinkOracle = await hre.ethers.getContract("ChainlinkOracle");
+  }
+
+  const defaultProxyAdmin = await artifacts.readArtifact(
+    "hardhat-deploy/solc_0.8/openzeppelin/proxy/transparent/ProxyAdmin.sol:ProxyAdmin",
+  );
+
+  await deploy("wstETHOneJumpChainlinkOracle", {
+    contract: "OneJumpOracle",
+    from: deployer,
+    log: true,
+    deterministicDeployment: false,
+    args: [wstETH, WETH, resilientOracle.address, chainlinkOracle.address],
+    proxy: {
+      owner: proxyOwnerAddress,
+      proxyContract: "OptimizedTransparentUpgradeableProxy",
+      viaAdminContract: {
+        name: "DefaultProxyAdmin",
+        artifact: defaultProxyAdmin,
+      },
+    },
+    skipIfAlreadyDeployed: true,
+  });
+
+  await deploy("weETHOneJumpChainlinkOracle", {
+    contract: "OneJumpOracle",
+    from: deployer,
+    log: true,
+    deterministicDeployment: false,
+    args: [weETH, WETH, resilientOracle.address, chainlinkOracle.address],
+    proxy: {
+      owner: proxyOwnerAddress,
+      proxyContract: "OptimizedTransparentUpgradeableProxy",
+      viaAdminContract: {
+        name: "DefaultProxyAdmin",
+        artifact: defaultProxyAdmin,
+      },
+    },
+    skipIfAlreadyDeployed: true,
+  });
+};
+
+func.skip = async () => hre.network.name !== "arbitrumone" && hre.network.name !== "arbitrumsepolia";
+func.tags = ["wstETH_weETH_OneJumpOracles_arbitrum"];
+export default func;