+++ deploy/5-deploy-bnb-lst-oracles.ts
@@ -4,7 +4,12 @@ import { HardhatRuntimeEnvironment } from "hardhat/types";
 
 import { ADDRESSES, assets } from "../helpers/deploymentConfig";
 
-const func: DeployFunction = async ({ getNamedAccounts, deployments, network }: HardhatRuntimeEnvironment) => {
+const func: DeployFunction = async ({
+  getNamedAccounts,
+  deployments,
+  network,
+  artifacts,
+}: HardhatRuntimeEnvironment) => {
   const { deploy } = deployments;
   const { deployer } = await getNamedAccounts();
 
@@ -14,7 +19,9 @@ const func: DeployFunction = async ({ getNamedAccounts, deployments, network }: 
   const { ankrBNB, stkBNB, BNBx, BNBxStakeManager, slisBNBStakeManager, stkBNBStakePool, slisBNB, wBETH } =
     ADDRESSES[network.name];
   const ETH = assets[network.name].find(asset => asset.token === "ETH");
-
+  const defaultProxyAdmin = await artifacts.readArtifact(
+    "hardhat-deploy/solc_0.8/openzeppelin/proxy/transparent/ProxyAdmin.sol:ProxyAdmin",
+  );
   await deploy("BNBxOracle", {
     from: deployer,
     log: true,
@@ -22,7 +29,11 @@ const func: DeployFunction = async ({ getNamedAccounts, deployments, network }: 
     args: [BNBxStakeManager, BNBx, oracle.address],
     proxy: {
       owner: proxyOwnerAddress,
-      proxyContract: "OptimizedTransparentProxy",
+      proxyContract: "OptimizedTransparentUpgradeableProxy",
+      viaAdminContract: {
+        name: "DefaultProxyAdmin",
+        artifact: defaultProxyAdmin,
+      },
     },
     skipIfAlreadyDeployed: true,
   });
@@ -34,7 +45,11 @@ const func: DeployFunction = async ({ getNamedAccounts, deployments, network }: 
     args: [slisBNBStakeManager, slisBNB, oracle.address],
     proxy: {
       owner: proxyOwnerAddress,
-      proxyContract: "OptimizedTransparentProxy",
+      proxyContract: "OptimizedTransparentUpgradeableProxy",
+      viaAdminContract: {
+        name: "DefaultProxyAdmin",
+        artifact: defaultProxyAdmin,
+      },
     },
     skipIfAlreadyDeployed: true,
   });
@@ -46,29 +61,16 @@ const func: DeployFunction = async ({ getNamedAccounts, deployments, network }: 
     args: [stkBNBStakePool, stkBNB, oracle.address],
     proxy: {
       owner: proxyOwnerAddress,
-      proxyContract: "OptimizedTransparentProxy",
+      proxyContract: "OptimizedTransparentUpgradeableProxy",
+      viaAdminContract: {
+        name: "DefaultProxyAdmin",
+        artifact: defaultProxyAdmin,
+      },
     },
     skipIfAlreadyDeployed: true,
   });
 
-  let ankrBNBAddress = ankrBNB;
-  if (!ankrBNB) {
-    // deploy MockAnkrBNB
-    await deploy("MockAnkrBNB", {
-      from: deployer,
-      log: true,
-      autoMine: true, // speed up deployment on local network (ganache, hardhat), no effect on live networks
-      skipIfAlreadyDeployed: true,
-      args: ["Ankr Staked BNB ", "ankrBNB", "18"],
-    });
-
-    const ankrBNBContract = await ethers.getContract("MockAnkrBNB");
-    ankrBNBAddress = ankrBNBContract.address;
-
-    if ((await ankrBNBContract.owner()) === deployer) {
-      await ankrBNBContract.transferOwnership(proxyOwnerAddress);
-    }
-  }
+  const ankrBNBAddress = ankrBNB || (await ethers.getContract("MockAnkrBNB")).address;
 
   await deploy("AnkrBNBOracle", {
     from: deployer,
@@ -77,29 +79,16 @@ const func: DeployFunction = async ({ getNamedAccounts, deployments, network }: 
     args: [ankrBNBAddress, oracle.address],
     proxy: {
       owner: proxyOwnerAddress,
-      proxyContract: "OptimizedTransparentProxy",
+      proxyContract: "OptimizedTransparentUpgradeableProxy",
+      viaAdminContract: {
+        name: "DefaultProxyAdmin",
+        artifact: defaultProxyAdmin,
+      },
     },
     skipIfAlreadyDeployed: true,
   });
 
-  let wBETHAddress = wBETH;
-  if (!wBETH) {
-    // deploy MockWBETH
-    await deploy("MockWBETH", {
-      from: deployer,
-      log: true,
-      autoMine: true, // speed up deployment on local network (ganache, hardhat), no effect on live networks
-      skipIfAlreadyDeployed: true,
-      args: ["Wrapped Binance Beacon ETH", "wBETH", "18"],
-    });
-
-    const wBETHContract = await ethers.getContract("MockWBETH");
-    wBETHAddress = wBETHContract.address;
-
-    if ((await wBETHContract.owner()) === deployer) {
-      await wBETHContract.transferOwnership(proxyOwnerAddress);
-    }
-  }
+  const wBETHAddress = wBETH || (await ethers.getContract("MockWBETH")).address;
 
   await deploy("WBETHOracle", {
     from: deployer,
@@ -108,7 +97,11 @@ const func: DeployFunction = async ({ getNamedAccounts, deployments, network }: 
     args: [wBETHAddress, ETH?.address, oracle.address],
     proxy: {
       owner: proxyOwnerAddress,
-      proxyContract: "OptimizedTransparentProxy",
+      proxyContract: "OptimizedTransparentUpgradeableProxy",
+      viaAdminContract: {
+        name: "DefaultProxyAdmin",
+        artifact: defaultProxyAdmin,
+      },
     },
     skipIfAlreadyDeployed: true,
   });